+++ src/app/services/horostorage/horostorage.service.ts
@@ -4,6 +4,7 @@ import { DirectionMethod } from 'src/app/process/enum/direction-method';
 import { ArcToDateMethod } from 'src/app/process/enum/arc-to-date-method';
 import { ProfectionArcToDateMethod } from 'src/app/process/enum/profection-arc-to-date-method';
 import { DailyDirectionMethod } from 'src/app/process/enum/daily-direction-method';
+import { SecondaryProgressionMethod } from 'src/app/process/enum/secondary-progression-method';
 import { DeepReadonly } from 'src/app/type/interface/deep-readonly';
 import {
   HoroRequest,
@@ -142,6 +143,7 @@ export class HoroStorageService {
         arc_to_date_method: ArcToDateMethod.DegreePerYear,
         profection_arc_to_date_method: ProfectionArcToDateMethod.TrueSolarArc,
         daily_direction_method: DailyDirectionMethod.SemiArcZodiacal,
+        secondary_progression_method: SecondaryProgressionMethod.DegreePerYear,
       });
     }
   }